Broken areas refactoring
Move all files to one folder
Add Readme
