+++ screenqual/filter/broken_areas/README.md
@@ -1,0 +1,36 @@
+Базовый класс и реализиции детекторов для нахождения недогрузившихся областей на скриншотах(картинки, видео). 
+
+##Broken images
+###Основная идея алгоритма
+Данный алгоритм находит недогрузившиеся картинки на скриншотах. Алгоритм находит только картинки, которые выглядят как одноцветные области(прямоугольники или прямоугольники с закругленными концами).
+Алгоритм не рассчитан на нахождение недогрузившихся картинок, для которых визуально нет их границы на скриншоте (например остается только белая область вместо картинки).
+ 
+Для нахождения данной аномалии на скриншоте находятся все "подозрительные" области, которые могли бы быть картинками и проверяется условие, что эти области одноцветные. 
+
+###Границы применимости
+Работает на десктопных, подокументных, мобильных скриншотах Яндекс и Google. Возможно так же работает и для скриншотов других сайтов, но данная функциональность не проверялась авторами. 
+
+###Известный проблемы
+Наблюдается ложноположительное срабатывание для ряда скриншотов, где содержится текст на цветном фоне(текст подцеркнут). Алгоритм может распознать часть фона как битую область.
+
+###Описание алгоритма
+1. Бинаризуем изображение и применяем морфологии, чтобы избавиться от артефактов. 
+2. Находим существующие контуры на изображении.
+3. Перебираем все контуры и каждый из них проверяем, является ли он прямоугольником(при аппроксимации в контуре остается 4 точки) или прямоугольником с закругленными краями(находим bounding box для контура и смотрим на соотношение площади контура к площади bounding box, если оно ниже, задаваемого в алгоритме порога, то считаем контур прямоугольников с закругленными краями)
+4. Проверяем найденные на предыдущем шаге контуры на выполнимость условия: все точки внутри контура одного цвета
+5. Если 4 выполняется хотя бы для одного контура, то считается, что скриншот с аномалией. 
+
+##Broken Videos
+Данный алгоритм находит недогрузившиеся видео на скриншотах.
+
+Для нахождения недогруженных видео вначале собирается тренировочное множество на основе которого тренируется простой классификатор, основыннай на гистограммах. Далее на основе данного классификатора прнимается решение, является ли скриншот битым.
+
+###Границы применимости
+Работает на десктопных, подокументных, мобильных скриншотах Яндекс и сатйами для которых недогруженные видео будут выглядеть похожим образом. Алгоритм не проверялся для других сайтов и требует от пользователя самостоятельного сбора тренировочного множества для работы с другими сайтами. 
+
+
+###Описание Алгоритма
+1. Найти на данном скриншоте "подозрительные" под видео области. Используется идея, аналогичная Broken images. 
+2. Для каждой области посчитать ее гистограмму(в качестве дефолтных значений считаем, что для нашей гистограммы есть 26 бинов).
+3. Произвести сравнение гистограммы с эталонной гистограммой, вычисленной на основе тренировачного множества. Для сравнения используется евклидова метрика. 
+4. По порогу определяется, является ли данная область недогруженным видео.
