+++ screenqual/filter/broken_areas/broken_areas_analyser.py
